refactor: remove adapter from naming
