+++ src/libecalc/domain/process/core/stream/mixing.py
@@ -12,7 +12,7 @@ from libecalc.domain.process.core.stream.exceptions import (
     ZeroTotalMassRateException,
 )
 from libecalc.domain.process.core.stream.stream import Stream
-from libecalc.domain.process.core.stream.thermo_system_adapter import NeqSimThermoSystem
+from libecalc.domain.process.core.stream.thermo_system import NeqSimThermoSystem
 
 
 class StreamMixingStrategy(Protocol):
