+++ src/libecalc/domain/process/core/stream/stream.py
@@ -6,7 +6,7 @@ from functools import cached_property
 from libecalc.common.units import UnitConstants
 from libecalc.domain.process.core.stream.conditions import ProcessConditions
 from libecalc.domain.process.core.stream.exceptions import NegativeMassRateException
-from libecalc.domain.process.core.stream.thermo_system_adapter import ThermoSystemInterface
+from libecalc.domain.process.core.stream.thermo_system import ThermoSystemInterface
 
 
 @dataclass(frozen=True)
