+++ tests/libecalc/domain/core/stream/test_stream.py
@@ -7,7 +7,7 @@ from libecalc.common.units import UnitConstants
 from libecalc.domain.process.core.stream.conditions import ProcessConditions
 from libecalc.domain.process.core.stream.exceptions import NegativeMassRateException
 from libecalc.domain.process.core.stream.stream import Stream
-from libecalc.domain.process.core.stream.thermo_system_adapter import NeqSimThermoSystem, ThermoSystemInterface
+from libecalc.domain.process.core.stream.thermo_system import NeqSimThermoSystem, ThermoSystemInterface
 
 
 @pytest.fixture
